+++ src/main/java/com/bobmowzie/mowziesmobs/server/ai/EntityAIBarakoayaTrade.java
@@ -1,16 +1,15 @@
 package com.bobmowzie.mowziesmobs.server.ai;
 
-import com.bobmowzie.mowziesmobs.server.entity.barakoa.EntityBarakoaya;
+import com.bobmowzie.mowziesmobs.server.entity.barakoa.EntityBarakoaVillager;
 import net.minecraft.entity.ai.goal.Goal;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.inventory.container.Container;
 
 import java.util.EnumSet;
 
 public final class EntityAIBarakoayaTrade extends Goal {
-    private final EntityBarakoaya barakoaya;
+    private final EntityBarakoaVillager barakoaya;
 
-    public EntityAIBarakoayaTrade(EntityBarakoaya barakoaya) {
+    public EntityAIBarakoayaTrade(EntityBarakoaVillager barakoaya) {
         this.barakoaya = barakoaya;
         setMutexFlags(EnumSet.of(Flag.MOVE, Flag.JUMP, Flag.TARGET));
     }
