+++ src/main/java/com/bobmowzie/mowziesmobs/server/ai/EntityAIBarakoayaTradeLook.java
@@ -1,15 +1,15 @@
 package com.bobmowzie.mowziesmobs.server.ai;
 
-import com.bobmowzie.mowziesmobs.server.entity.barakoa.EntityBarakoaya;
+import com.bobmowzie.mowziesmobs.server.entity.barakoa.EntityBarakoaVillager;
 import net.minecraft.entity.ai.goal.LookAtGoal;
 import net.minecraft.entity.player.PlayerEntity;
 
 import java.util.EnumSet;
 
 public class EntityAIBarakoayaTradeLook extends LookAtGoal {
-    private final EntityBarakoaya barakoaya;
+    private final EntityBarakoaVillager barakoaya;
 
-    public EntityAIBarakoayaTradeLook(EntityBarakoaya barakoaya) {
+    public EntityAIBarakoayaTradeLook(EntityBarakoaVillager barakoaya) {
         super(barakoaya, PlayerEntity.class, 8);
         this.barakoaya = barakoaya;
         setMutexFlags(EnumSet.of(Flag.LOOK));
