+++ src/main/java/com/bobmowzie/mowziesmobs/server/world/spawn/SpawnHandler.java
@@ -52,7 +52,7 @@ public class SpawnHandler {
             EntitySpawnPlacementRegistry.register(EntityHandler.BARAKOANA, mmSpawn, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, MowzieEntity::spawnPredicate);
             EntitySpawnPlacementRegistry.register(EntityHandler.NAGA, EntitySpawnPlacementRegistry.PlacementType.NO_RESTRICTIONS, Heightmap.Type.MOTION_BLOCKING, MowzieEntity::spawnPredicate);
             EntitySpawnPlacementRegistry.register(EntityHandler.GROTTOL, mmSpawn, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, MowzieEntity::spawnPredicate);
-            EntitySpawnPlacementRegistry.register(EntityHandler.BARAKOA_SUNBLOCKER, mmSpawn, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, MowzieEntity::spawnPredicate);
+            EntitySpawnPlacementRegistry.register(EntityHandler.BARAKOAYA, mmSpawn, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, MowzieEntity::spawnPredicate);
         }
     }
